test: activate and improve genconfig command tests
- Activated genconfig tests that were marked as "toremove"
- Updated genconfig to support filesystem injection for testing
- Added comprehensive test coverage for all genconfig functionality
- Added documentation for command reference and testing gaps
- Identified areas where tests were passing but commands were broken

This addresses the issue where we thought commands were working because
tests passed, but the commands were actually broken (like init and fill).
