+++ pkg/commands/genconfig/genconfig.go
@@ -2,11 +2,11 @@ package genconfig
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
 	"strings"
 
 	"github.com/arthur-debert/dodot/pkg/config"
+	"github.com/arthur-debert/dodot/pkg/filesystem"
 	"github.com/arthur-debert/dodot/pkg/logging"
 	"github.com/arthur-debert/dodot/pkg/types"
 )
@@ -16,11 +16,18 @@ type GenConfigOptions struct {
 	DotfilesRoot string
 	PackNames    []string
 	Write        bool
+	FileSystem   types.FS // Optional filesystem for testing
 }
 
 // GenConfig outputs or writes the default configuration
 func GenConfig(opts GenConfigOptions) (*types.GenConfigResult, error) {
 	logger := logging.GetLogger("commands.genconfig")
+
+	// Use provided filesystem or default
+	fs := opts.FileSystem
+	if fs == nil {
+		fs = filesystem.NewOS()
+	}
 
 	// Get the user defaults content from the config package
 	userDefaultsContent := config.GetUserDefaultsContent()
@@ -61,18 +68,18 @@ func GenConfig(opts GenConfigOptions) (*types.GenConfigResult, error) {
 	for _, targetPath := range targetPaths {
 		// Ensure directory exists
 		dir := filepath.Dir(targetPath)
-		if err := os.MkdirAll(dir, 0755); err != nil {
+		if err := fs.MkdirAll(dir, 0755); err != nil {
 			return result, fmt.Errorf("failed to create directory %s: %w", dir, err)
 		}
 
 		// Check if file already exists
-		if _, err := os.Stat(targetPath); err == nil {
+		if _, err := fs.Stat(targetPath); err == nil {
 			logger.Warn().Str("path", targetPath).Msg("Config file already exists, skipping")
 			continue
 		}
 
 		// Write the file
-		if err := os.WriteFile(targetPath, []byte(result.ConfigContent), 0644); err != nil {
+		if err := fs.WriteFile(targetPath, []byte(result.ConfigContent), 0644); err != nil {
 			return result, fmt.Errorf("failed to write config to %s: %w", targetPath, err)
 		}
